ai-service/requirements.txt
flask
flask-cors
scikit-learn
numpy
pandas
PyPDF2
pdfplumber
python-dotenv
python-docx
textblob
